+++ Source/ACE/Entity/Creature.cs
@@ -259,6 +259,10 @@ namespace ACE.Entity
             }
         }
 
+        /// <summary>
+        /// This method checks to make sure we have a casting device equipped and if so, it sets
+        /// the motion state and sends the messages to switch us to spellcasting state.   Og II
+        /// </summary>
         public void HandleSwitchToMagicCombatMode()
         {
             HeldItem mEquipedWand = Children.Find(s => s.EquipMask == EquipMask.Held);
@@ -273,6 +277,11 @@ namespace ACE.Entity
                 log.InfoFormat("Changing combat mode for {0} - could not locate a wielded magic caster", Guid);
         }
 
+        /// <summary>
+        /// This method is called if we unwield missle ammo.   It will check to see if I have arrows wielded
+        /// send the message to "hide" the arrow.
+        /// </summary>
+        /// <param name="oldCombatMode"></param>
         public void HandleUnloadMissileAmmo(CombatMode oldCombatMode)
         {
             // Before I can switch to any non missile mode, do I have missile ammo I need to remove?
@@ -292,6 +301,12 @@ namespace ACE.Entity
             }
         }
 
+        /// <summary>
+        /// This method sets us into peace mode.   It checks our current state to see if we have missle ammo equipped
+        /// it will make the call to hid the "ammo" as we switch to peace mode.   It will then send the message switch our stance. Og II
+        /// </summary>
+        /// <param name="oldCombatMode"></param>
+        /// <param name="isAutonomous"></param>
         public void HandleSwitchToPeaceMode(CombatMode oldCombatMode, bool isAutonomous = false)
         {
             HandleUnloadMissileAmmo(oldCombatMode);
@@ -303,7 +318,10 @@ namespace ACE.Entity
             UniversalMotion mm = new UniversalMotion(MotionStance.Standing);
             mm.MovementData.CurrentStyle = (ushort)MotionStance.Standing;
             SetMotionState(this, mm);
+            var mEquipedAmmo = WieldedObjects.First(s => s.Value.CurrentWieldedLocation == EquipMask.MissileAmmo).Value;
             CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange, new GameMessagePrivateUpdatePropertyInt(Sequences, PropertyInt.CombatMode, (uint)CombatMode.NonCombat));
+            if (mEquipedAmmo != null)
+                CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectGhostRange, new GameMessagePickupEvent(mEquipedAmmo));
         }
 
         public void HandleSwitchToMissileCombatMode(ActionChain combatModeChain)
@@ -317,7 +335,9 @@ namespace ACE.Entity
                     log.InfoFormat("Changing combat mode for {0} - could not locate wielded weapon {1}", Guid, mEquipedMissile.Guid);
                     return;
                 }
-                HeldItem mEquipedAmmo = Children.Find(s => s.EquipMask == EquipMask.MissileAmmo);
+
+                var mEquipedAmmo = WieldedObjects.First(s => s.Value.CurrentWieldedLocation == EquipMask.MissileAmmo).Value;
+
                 MotionStance ms;
 
                 if (missileWeapon.DefaultCombatStyle != null)
@@ -337,7 +357,6 @@ namespace ACE.Entity
                 }
                 else
                 {
-                    WorldObject ammo = GetInventoryItem(new ObjectGuid(mEquipedAmmo.Guid));
                     CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange, new GameMessageUpdatePosition(this));
                     SetMotionState(this, mm);
                     mm.MovementData.ForwardCommand = (ushort)MotionCommand.Reload;
@@ -350,7 +369,7 @@ namespace ACE.Entity
                     SetMotionState(this, mm);
                     // FIXME: (Og II)<this is a hack for now to be removed. Need to pull delay from dat file
                     combatModeChain.AddDelaySeconds(0.40);
-                    combatModeChain.AddAction(this, () => CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange, new GameMessageParentEvent(this, ammo, 1, 1)));
+                    combatModeChain.AddAction(this, () => CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange, new GameMessageParentEvent(this, mEquipedAmmo, 1, 1)));
                     // CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange, new GameMessageParentEvent(this, ammo, 1, 1)); // used for debugging
                 }
                 CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange, new GameMessagePrivateUpdatePropertyInt(Sequences, PropertyInt.CombatMode, (uint)CombatMode.Missile));