+++ changelog.md
@@ -1,9 +1,16 @@
 # ACEmulator Change Log
 
 
-### 2017-09-21
+### 2017-09-25
 [Og II]
-* WIP persist wielded items.   I am posting this looking for some help.
+* Implemented persisting of wielded items.   
+* Changed inventory to manage a list of world objects instead of ace objects.   This was needed for sequence issues.
+* fixed combat animation that had been broken for a while due to prior refactoring.
+* fixed moving inventory from container to container with persistances which was broken due to prior refactoring.
+* implemented the response message to send object description.
+* cleaned up so code, aligned enums.
+* TODO.   I have a dirty flag bug that causes issues when you do container to container moves and save.   
+* Not sure if you want to accept this with a known bug - it has changed a lot and is probably holding others up.
 
 ### 2017-09-20
 [Ripley]